+++ test/api/lighthouse-service.js
@@ -5,6 +5,16 @@ import { serviceUrl } from "../fixtures.js";
 export const lighthouseService = {
   lighthouseUrl: serviceUrl,
 
+  async authenticate(user) {
+    const response = await axios.post(`${this.lighthouseUrl}/api/users/authenticate`, user);
+    axios.defaults.headers.common.Authorization = `Bearer ${  response.data.token}`;
+    return response.data;
+  },
+
+  async clearAuth() {
+    axios.defaults.headers.common.Authorization = "";
+  },
+
   async createUser(user) {
     const res = await axios.post(`${this.lighthouseUrl}/api/users`, user);
     return res.data;