+++ test/api/user-api-test.js
@@ -7,12 +7,16 @@ const users = new Array(testUsers.length);
 
 suite("User API tests", () => {
   setup(async () => {
-    // db.init("mongo");
+    lighthouseService.clearAuth();
+    await lighthouseService.createUser(maggie);
+    await lighthouseService.authenticate(maggie);
     await lighthouseService.deleteAllUsers();
     for (let i = 0; i < testUsers.length; i += 1) {
       // eslint-disable-next-line no-await-in-loop
     users[i] = await lighthouseService.createUser(testUsers[i]);
-    }
+    };
+    await lighthouseService.createUser(maggie);
+    await lighthouseService.authenticate(maggie);
   });
   teardown(async () => {
   });
@@ -25,10 +29,12 @@ suite("User API tests", () => {
 
   test("delete all users", async () => {
     let returnedUsers = await lighthouseService.getAllUsers();
-    assert.equal(returnedUsers.length, 3);
+    assert.equal(returnedUsers.length, 4);
     await lighthouseService.deleteAllUsers();
+    await lighthouseService.createUser(maggie);
+    await lighthouseService.authenticate(maggie);
     returnedUsers = await lighthouseService.getAllUsers();
-    assert.equal(returnedUsers.length, 0);
+    assert.equal(returnedUsers.length, 1);
   });
 
   test("get a user - success", async () => {
@@ -48,6 +54,8 @@ suite("User API tests", () => {
 
   test("get a user - deleted user", async () => {
     await lighthouseService.deleteAllUsers();
+    await lighthouseService.createUser(maggie);
+    await lighthouseService.authenticate(maggie);
     try {
       const returnedUser = await lighthouseService.getUser(users[0]._id);
       assert.fail("Should not return a response");